+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.4.0] - 24-06-11
+
+### Added
+
+- Added support for `SourceDescription.poster` for Android & iOS.
+- Made the intervals for the forward and backward skip buttons on the iOS lockScreen configurable.
+- Added preferredPeakBitRate and preferredMaximumResolution to ABRConfiguration for iOS
+
 ## [7.3.0] - 24-06-03
 
 ### Changed